server.js and product.controller.js modified

+++ backend/controllers/product.controller.js
@@ -42,6 +42,7 @@ export const updateProduct = async (req, res) => {
         const updteProduct = await Product.findByIdAndUpdate(id, product, { new: true });
         res.status(200).json({ success: true, data: updteProduct });
     } catch (error) {
+        console.log("Error in updating products", error.message);
         res.status(500).json({ success: false, message: "Server Error" });
     }
 };
@@ -49,11 +50,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid product Id" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product Deleted" });
     } catch (error) {
         console.log("Error in delteing products", error.message);
-        res.status(400).json({ success:false, message: "Product not found" });
+        res.status(500).json({ success:false, message: "Server Error" });
     }
 };
